--- a/integration_tests/derived_types_13c.f90
+++ b/integration_tests/derived_types_13c.f90
@@ -1,0 +1,4 @@
+module debug_3
+    use debug_2, only: sub_1
+    implicit none
+end module debug_3