--- a/integration_tests/enum_02.f90
+++ b/integration_tests/enum_02.f90
@@ -1,0 +1,7 @@
+program enum_02
+    use enum_02_module, only: compiler_enum
+    implicit none
+
+    print *, compiler_enum
+
+end program enum_02