--- a/integration_tests/enum_02_module.f90
+++ b/integration_tests/enum_02_module.f90
@@ -1,0 +1,10 @@
+module enum_02_module
+    implicit none
+    enum, bind(c)
+        enumerator :: red, blue
+        enumerator yellow
+    end enum
+
+    integer, parameter :: compiler_enum = kind(red)
+
+end module enum_02_module