--- a/integration_tests/intrinsics_47.f90
+++ b/integration_tests/intrinsics_47.f90
@@ -1,0 +1,5 @@
+program intrinsics_47
+  logical l
+  l = all([.true., .true.])
+  if (.not.(l)) error stop
+end program intrinsics_47