--- a/tests/modules11.f90
+++ b/tests/modules11.f90
@@ -1,0 +1,5 @@
+module fpm_cmd_install
+use fpm, only : build_model
+implicit none
+
+end module fpm_cmd_install