Continues https://github.com/lfortran/lfortran/pull/1153

Commits to be added in `main`,

- [x] e767b428169b324e2f52d366aece8a79813bfe00 - https://github.com/lfortran/lfortran/pull/1273 
- [x] 81c32589527981e45aac6a66ac89348c130dbb18 - https://github.com/lfortran/lfortran/pull/1273
- [x] 12b1f899b4268df0f020165cb3184de2eb9414c8 - https://github.com/lfortran/lfortran/pull/1269
- [x] bf7aa415ec510106caa4b91c723baf0e1aadb88a - https://github.com/lfortran/lfortran/pull/1269
- [x] 056b4f2125a115bccec14b48d8b25c22451d8dbe - https://github.com/lfortran/lfortran/pull/1270
- [x] 08b8658d92aeff18053e485cb4cbfe293ec6d975 - https://github.com/lfortran/lfortran/pull/1270
- [x] 653267b9aeacad1adf72f12fcbb2af8655ee4965 - https://github.com/lfortran/lfortran/pull/1271
- [x] 19fe4305ce5321e7e7d218871eb5567a66fe3e26 - https://github.com/lfortran/lfortran/pull/1295
- [x] f03d356ba9a987fd59eb0ff721063c6385cde01a - https://github.com/lfortran/lfortran/pull/1272 
- [x] cc5dff2fe77d847d9df5c633d5cd3229ecdc33cc - https://github.com/lfortran/lfortran/pull/1272 
- [x] 4b6a482113e74da21bc5cacee154e0be22108f8d - https://github.com/lfortran/lfortran/pull/1272
- [x] edb65d4d1b650efea1629c6299714d26c7c43b57 - https://github.com/lfortran/lfortran/pull/1272
- [x] cecb51d0b9907c546a72e642f7301ee3b0d16018 - https://github.com/lfortran/lfortran/pull/1272
- [x] 8ccc71dd3184908d4ba946a230b4eba62d745983 - https://github.com/lfortran/lfortran/pull/1288 
- [x] 908b9f3cf3da9cddc4b514737896d9efc1c6c4d3 - https://github.com/lfortran/lfortran/pull/1288
- [x] 47eed37a39ff572eb6b69cd0862ef75adcecec14 - https://github.com/lfortran/lfortran/pull/1288
- [x] 6ea8b73020669248ad4f5a311e513876ddc900ae - https://github.com/lfortran/lfortran/pull/1276
- [x] feffbd9240120adfa063aa1adb532569c1acde9c - https://github.com/lfortran/lfortran/pull/1277 
- [x] 23e0921c868913559da4e655b094bb24750fe27e - https://github.com/lfortran/lfortran/pull/1296
- [x] 9fbd653edb1bd5979fae9553591cbd4343914fbd - https://github.com/lfortran/lfortran/pull/1289
- [x] 9116df333caa7135a0ef4ed56732418a5c7d50d8 - https://github.com/lfortran/lfortran/pull/1283
- [x] b6133a9c258e62f2e775da35621385f465a3430c - https://github.com/lfortran/lfortran/pull/1288 
- [x] cb0b767f7c724153aa4a2f9ac591e787a1f083e1 - https://github.com/lfortran/lfortran/pull/1296
- [x] 90817bd43f10b5f024e5af74dac49214f76e98ce - https://github.com/lfortran/lfortran/pull/1276
- [x] 579435d1f48081914bff0d361d25a04178013f41 - https://github.com/lfortran/lfortran/pull/1279 
- [x] 4bc78805431f0f95e570b4a9e429980022f5f42a - https://github.com/lfortran/lfortran/pull/1280 
- [x] 7ff56529e93b06cf38bcc3c5ed5672a9581f2f82 - https://github.com/lfortran/lfortran/pull/1280 
- [x] 7bb475f5b8f7cb5658b8bf1ce89a2b5e11801fa6 - https://github.com/lfortran/lfortran/pull/1277
- [x] a13c6703eb6cfeda7291ad5747feaa7979f9490f - https://github.com/lfortran/lfortran/pull/1288
- [x] 0e578dfa12bdedc8ed467b2f701d686659e2e58a - https://github.com/lfortran/lfortran/pull/1290 
- [x] 8728172dcb3471bbb437e5d3d26a99ed61b56507 - https://github.com/lfortran/lfortran/pull/1278
- [x] 5f583b8fb471fa296a993a5cad897b1913255232 - https://github.com/lfortran/lfortran/pull/1290
- [x] 4474612885bca5879c1614b7029a450207c60981 - https://github.com/lfortran/lfortran/pull/1281 
- [x] 9b4c5c6e0daa762275df38783d9ab561989e2466 - https://github.com/lfortran/lfortran/pull/1279
- [x] 6df1d0c2665b379d222030ba65d2d978dff89c0b - https://github.com/lfortran/lfortran/pull/1281 
- [x] 282c93cf249aceb9f2d12c4266c500d04c67e467 - https://github.com/lfortran/lfortran/pull/1282 

@Thirumalai-Shaktivel @Smit-create Please continue your sprint here. Also please do not rebase on top of main branch. I need the commits to cherry-pick in future, so **please do not change the git history of the PR**. Just add new commits and push.

Progress:

```bash
(lf) 16:13:11:~/lfortran_project/fpm % FPM_FC="lfortran" fpm build --flag="--cpp"
filesystem_utilities.c                 done.
fpm_strings.f90                        done.
fpm_backend_console.f90                done.
iscygpty.c                             done.
isatty.c                               done.
constants.f90                          done.
version.f90                            done.
M_CLI2.F90                             done.
error.f90                              done.
error.f90                              done.
datetime.f90                           done.
verify.f90                             done.
fpm_environment.f90                    done.
fpm_os.F90                             done.
versioning.f90                         done.
convert.f90                            done.
fpm_filesystem.F90                     done.
utils.f90                              done.
fpm_command_line.f90                   done.
installer.f90                          done.
git.f90                                done.
value.f90                              done.
sort.f90                               done.
keyval.f90                             done.
base.f90                               done.
vector.f90                             done.
structure.f90                          done.
table.f90                              done.
array.f90                              done.
type.f90                               done.
ser.f90                                done.
tokenizer.f90                          done.
merge.f90                              done.
keyval.f90                             done.
character.f90                          done.
table.f90                              done.
array.f90                              done.
de.f90                                 done.
build.f90                              done.
tomlf.f90                              done.
toml.f90                               done.
profiles.f90                           done.
install.f90                            done.
preprocess.f90                         done.
build.f90                              done.
dependency.f90                         done.
library.f90                            done.
executable.f90                         done.
test.f90                               done.
example.f90                            done.
package.f90                            done.
manifest.f90                           done.
new.f90                                done.
fpm_compiler.f90                       done.
dependency.f90                         done.
fpm_model.f90                          done.
update.f90                             done.
fpm_source_parsing.f90                 done.
fpm_targets.f90                        done.
fpm_backend_output.f90                 done.
fpm_sources.f90                        done.
fpm_backend.F90                        done.
fpm.f90                                done.
install.f90                            done.
libfpm.a                               done.
main.f90                               failed.
[ 98%] Compiling...
ASR verify pass error: ASR verify: Function get_command_line_settings doesn't depend on any but is found in its dependency list.
Internal Compiler Error: Unhandled exception
Traceback (most recent call last):
  File "/Users/czgdp1807/lfortran_project/lfortran/src/bin/lfortran.cpp", line 1907
    return compile_to_object_file(arg_file, outfile, false,
  File "/Users/czgdp1807/lfortran_project/lfortran/src/bin/lfortran.cpp", line 865
    res = fe.get_llvm3(*asr, lpm, diagnostics, infile);
  File "/Users/czgdp1807/lfortran_project/lfortran/src/lfortran/fortran_evaluator.cpp", line 342
    compiler_options, run_fn, infile);
  File "/Users/czgdp1807/lfortran_project/lfortran/src/libasr/codegen/asr_to_llvm.cpp", line 6794
    pass_manager.apply_passes(al, &asr, pass_options, diagnostics);
  File "/Users/czgdp1807/lfortran_project/lfortran/src/libasr/pass/pass_manager.h", line 209
    _apply_passes(al, asr, _passes, pass_options, diagnostics);
  File "/Users/czgdp1807/lfortran_project/lfortran/src/libasr/pass/pass_manager.h", line 106
    throw LCompilersException("Verify failed");
LCompilersException: Verify failed
<ERROR> Compilation failed for object " app_main.f90.o "
<ERROR>stopping due to failed compilation
STOP 1
```