@czgdp1807 2023-02-02
@Thirumalai-Shaktivel Please feel free continue the sprint. I will be out of office tomorrow. I have progressed till 88%.