@Thirumalai-Shaktivel 2023-02-07
Next example on which I was working on:
```fortran
module module_1
    implicit none
contains
subroutine sub_1(a, b, c, d)
    integer, intent(in) :: a
    integer, intent(in), optional :: b
    integer, intent(in), optional :: c
    integer, intent(in) :: d
end subroutine sub_1

end module module_1

module module_2
    use module_1, only: sub_1
contains

    subroutine sub_2()
        call sub_1(1, b=1, c=1, d=2)
        ! call sub_1(1, d=2)
    end subroutine sub_2
end module module_2
```