@czgdp1807 2023-02-08
So compilation till ASR is complete. As you can see in the description of this PR, the error is in LLVM backend now. So, I would mark this sprint as complete. We will know cherry-pick commits and apply to the PR. Later today, I will update the description of this PR with list of commits (and assign to @Smit-create, @Thirumalai-Shaktivel and myself) to be merged into main.