@certik 2023-02-08
Yes, let's merge everything into main and setup a CI to compile all of fpm into ASR, to ensure this keeps working.