@certik 2023-02-13
I tested fpm version 9cb38b2a9f3c31d6617a48e757cd56556b0d05b5, LFortran in Release mode on Apple M1 Max:
```console
$ time FPM_FC="lfortran" fpm build --flag="--cpp"
filesystem_utilities.c                 done.
fpm_strings.f90                        done.
fpm_backend_console.f90                done.
iscygpty.c                             done.
isatty.c                               done.
constants.f90                          done.
version.f90                            done.
M_CLI2.f90                             done.
error.f90                              done.
error.f90                              done.
datetime.f90                           done.
verify.f90                             done.
fpm_environment.f90                    done.
fpm_os.F90                             done.
versioning.f90                         done.
convert.f90                            done.
fpm_filesystem.F90                     done.
utils.f90                              done.
fpm_command_line.f90                   done.
installer.f90                          done.
git.f90                                done.
value.f90                              done.
keyval.f90                             done.
base.f90                               done.
vector.f90                             done.
structure.f90                          done.
table.f90                              done.
array.f90                              done.
type.f90                               done.
ser.f90                                done.
tokenizer.f90                          done.
merge.f90                              done.
keyval.f90                             done.
character.f90                          done.
table.f90                              done.
array.f90                              done.
de.f90                                 done.
build.f90                              done.
tomlf.f90                              done.
toml.f90                               done.
profiles.f90                           done.
install.f90                            done.
preprocess.f90                         done.
build.f90                              done.
dependency.f90                         done.
library.f90                            done.
executable.f90                         done.
test.f90                               done.
example.f90                            done.
package.f90                            done.
manifest.f90                           done.
new.f90                                done.
fpm_compiler.f90                       done.
dependency.f90                         done.
fpm_model.f90                          done.
update.f90                             done.
fpm_source_parsing.f90                 done.
fpm_targets.f90                        done.
fpm_backend_output.f90                 done.
fpm_sources.f90                        done.
fpm_backend.F90                        done.
fpm.f90                                done.
install.f90                            done.
libfpm.a                               done.
main.f90                               compiling...
[ 96%] Compiling...sh: line 1: 11129 Segmentation fault: 11  lfortran -c app/main.f90 --cpp -J build/lfortran_BC003944AE0DF3D6 -Ibuild/lfortran_BC003944AE0DF3D6 -o build/lfortran_BC003944AE0DF3D6/fpm/app_main.f90.o > build/lfortran_BC003944main.f90                               failed.
[ 98%] Compiling...
<ERROR> Compilation failed for object " app_main.f90.o "
<ERROR>stopping due to failed compilation
STOP 1
FPM_FC="lfortran" fpm build --flag="--cpp"  1.02s user 0.27s system 92% cpu 1.402 total
```

And GFortran 11.3.0
```console
$ time fpm build
filesystem_utilities.c                 done.
fpm_strings.f90                        done.
fpm_backend_console.f90                done.
iscygpty.c                             done.
isatty.c                               done.
constants.f90                          done.
version.f90                            done.
M_CLI2.f90                             done.
error.f90                              done.
error.f90                              done.
datetime.f90                           done.
verify.f90                             done.
fpm_environment.f90                    done.
fpm_os.F90                             done.
versioning.f90                         done.
convert.f90                            done.
fpm_filesystem.F90                     done.
utils.f90                              done.
fpm_command_line.f90                   done.
installer.f90                          done.
git.f90                                done.
value.f90                              done.
keyval.f90                             done.
base.f90                               done.
vector.f90                             done.
structure.f90                          done.
table.f90                              done.
array.f90                              done.
type.f90                               done.
ser.f90                                done.
tokenizer.f90                          done.
merge.f90                              done.
keyval.f90                             done.
character.f90                          done.
table.f90                              done.
array.f90                              done.
de.f90                                 done.
build.f90                              done.
tomlf.f90                              done.
toml.f90                               done.
profiles.f90                           done.
install.f90                            done.
preprocess.f90                         done.
build.f90                              done.
dependency.f90                         done.
library.f90                            done.
executable.f90                         done.
test.f90                               done.
example.f90                            done.
package.f90                            done.
manifest.f90                           done.
new.f90                                done.
fpm_compiler.f90                       done.
dependency.f90                         done.
fpm_model.f90                          done.
update.f90                             done.
fpm_source_parsing.f90                 done.
fpm_targets.f90                        done.
fpm_backend_output.f90                 done.
fpm_sources.f90                        done.
fpm_backend.F90                        done.
fpm.f90                                done.
install.f90                            done.
libfpm.a                               done.
main.f90                               done.
fpm                                    done.
[100%] Project compiled successfully.
fpm build  6.36s user 0.81s system 95% cpu 7.530 total
```
So LFortran is 5.3x faster to compile, so far. We'll see how much extra the ASR->LLVM adds. It looks promising.